scripts: drgn: add get_anon script
This is a rough script that is not brilliantly implemented but does the job. It
can be invoked from another program to retrieve anon_vma state from the command
line, e.g.:

$ scripts/drgn/get_anon <pid> <addr>

I hope to implement a libdrgn version of this at some point.

Also update README.md to reflect.